Add new components for intervention CostBenefitAnalysis
Introduced CostBenefitAnalysis, FindTreatments, Header, HealthTracking, and TopTreatments components. Updated DFDAPage to include these components for a more modularized approach, enhancing maintainability and readability.
